Corrección switch

+++ js/tareaswitch.js
@@ -1,5 +1,5 @@
 //En un establecimiento de juegos existen cuatro salas: Arcade, Consolas, Rol, Realidad Virtual. Si un usuario paga 4 créditos puede acceder a todas, si paga 3 solo podrá acceder a las tres primeras, si paga 2 a las dos primeras y si paga 1 solo a la primera sala.
-let creditos = parseFloat(prompt("Ingrese la cantidad de creditos"));
+let creditos = parseInt(prompt("Ingrese la cantidad de creditos"));
 
 switch (true) {
   case creditos < 1:
